% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCropareaLandInG.R
\name{calcCropareaLandInG}
\alias{calcCropareaLandInG}
\title{calcCropareaLandInG}
\usage{
calcCropareaLandInG(
  sectoral = "kcr",
  physical = TRUE,
  cellular = FALSE,
  irrigation = FALSE,
  selectyears = "all",
  climatetype = "MRI-ESM2-0:ssp370"
)
}
\arguments{
\item{sectoral}{"kcr" MAgPIE items, and "lpj" LPJmL items}

\item{physical}{if TRUE the sum over all crops plus fallow land (of calcFallowLand)
agrees with the physical cropland of readLandInG(subtype = physical)}

\item{cellular}{if TRUE: calculates cellular crop area for all magpie croptypes.
Option FALSE is not (yet) available.}

\item{irrigation}{If true: cellular areas are returned separated
into irrigated and rainfed}

\item{selectyears}{extract certain years from the data}

\item{climatetype}{Climate scenario or historical baseline "GSWP3-W5E5:historical"
used to determine multiple cropping suitability}
}
\value{
MAgPIE object with cropareas
}
\description{
This function uses total physical area and
             crop-specific harvested area data from LandInG
             to calculate crop-specific physical and harvested
             areas considering special rules
             for the allocation of perennial and annual crops.
}
\author{
David Hoetten, Felicitas Beier
}
