% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFallow.R
\name{calcFallow}
\alias{calcFallow}
\title{calcFallow}
\usage{
calcFallow(cellular = FALSE)
}
\arguments{
\item{cellular}{TRUE for cellular outputs}
}
\value{
MAgPIE object containing fallow land in Mha
}
\description{
Returns fallow land calculated based on physical cropland extent
             and harvested area output of default croparea datasource.
             The resultant areas are scaled to match the land croparea land
             class of calcLanduseInitialisation.
}
\examples{
\dontrun{
calcOutput("Fallow")
}
}
\seealso{
\code{\link{readLandInG}}
}
\author{
David Hoetten, Felicitas Beier, Benjamin Bodirsky
}
