% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrassGPPmonthly.R
\name{calcGrassGPPmonthly}
\alias{calcGrassGPPmonthly}
\title{calcGrassGPPmonthly}
\usage{
calcGrassGPPmonthly(selectyears, lpjml, climatetype)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates monthly gross primary production (GPP) of grassland
             under irrigated and rainfed conditions based on LPJmL inputs.
}
\examples{
\dontrun{
calcOutput("GrassGPP", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
