% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrassGPPyearly.R
\name{calcGrassGPPyearly}
\alias{calcGrassGPPyearly}
\title{calcGrassGPPyearly}
\usage{
calcGrassGPPyearly(selectyears, lpjml, climatetype, season)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or historical baseline "GSWP3-W5E5:historical"}

\item{season}{"wholeYear":  grass GPP in the entire year (main + off season)
"mainSeason": grass GPPP in the crop-specific growing
              period of LPJmL (main season)}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates gross primary production (GPP) of grassland
             under irrigated and rainfed conditions based on LPJmL inputs.
}
\examples{
\dontrun{
calcOutput("GrassGPPyearly", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
