% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmLHarmonize.R
\name{calcLPJmLHarmonize}
\alias{calcLPJmLHarmonize}
\title{calcLPJmLHarmonize}
\usage{
calcLPJmLHarmonize(
  lpjmlversion = "lpjml5.10.0-m4",
  climatetype = "MRI-ESM2-0:ssp245",
  subtype = "pnv:soilc",
  subdata = NULL,
  monthly = FALSE
)
}
\arguments{
\item{lpjmlversion}{Switch between LPJmL versions (including addons for further version specification)}

\item{climatetype}{Switch between different climate scenarios}

\item{subtype}{Switch between different LPJmL output variables as specified in readLPJmL}

\item{subdata}{Selection of subitems of object.
This argument can be used to split up the data in smaller objects
where only a sub-set of the data is needed or
for better handling where otherwise memory issues would occur due to the object size.}

\item{monthly}{Boolean indicating whether monthly data (if available)
should be returned (TRUE) or whether it should be aggregated to yearly values (FALSE).
If FALSE: Need to indicate aggregation method, e.g.,
"sum" (for fluxes), "mean" (for stocks/pools);
Argument should be separated by ":" (e.g., "FALSE:sum")}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Handle harmonization of LPJmL data
}
\examples{
\dontrun{
calcOutput("LPJmLHarmonize", subtype = "pnv:soilc", aggregate = FALSE)
}
}
\seealso{
[toolLPJmLHarmonize()]
}
\author{
Kristine Karstens, Felicitas Beier
}
