% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLUH3.R
\name{calcLUH3}
\alias{calcLUH3}
\title{calcLUH3}
\usage{
calcLUH3(
  landuseTypes = "magpie",
  irrigation = FALSE,
  cellular = FALSE,
  yrs = seq(1965, 2020, 5)
)
}
\arguments{
\item{landuseTypes}{magpie: magpie landuse classes,
LUH3: original landuse classes}

\item{irrigation}{if true: areas are returned separated by irrigated and rainfed,
if false: total areas (irrigated + rainfed)}

\item{cellular}{if true: dataset is returned on 0.5 degree resolution,
if false: return country-level data}

\item{yrs}{years to be returned}
}
\value{
magpie object with land data in Mha
}
\description{
Prepares the LUH3 historic landuse-dataset for usage in MAgPIE.
}
\examples{
\dontrun{
calcOutput("LUH3")
}
}
\seealso{
[calcLanduseInitialisation()]
}
\author{
Wanderson Costa, Pascal Sauer, Miodrag Stevanovic, Alexandre Koberle
}
