% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLUH3flood.R
\name{calcLUH3flood}
\alias{calcLUH3flood}
\title{calcLUH3flood}
\usage{
calcLUH3flood(cellular = FALSE, yrs = seq(1965, 2020, 5))
}
\arguments{
\item{cellular}{if true: dataset is returned on 0.5 degree resolution,
if false: return country-level data}

\item{yrs}{years to be returned}
}
\value{
magpie object with flood data in Mha
}
\description{
Prepares the LUH3 historic flood data for usage in MAgPIE,
by calculating the share of c3ann with c3ann area in Mha to obtain flood area in Mha.
}
\examples{
\dontrun{
calcOutput("LUH3flood")
}
}
\seealso{
[calcLanduseInitialisation()]
}
\author{
Wanderson Costa, Pascal Sauer, Miodrag Stevanovic, Alexandre Koberle
}
