% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLUHTotalLandArea.R
\name{calcLUHTotalLandArea}
\alias{calcLUHTotalLandArea}
\title{calcLUHTotalLandArea}
\usage{
calcLUHTotalLandArea(datasource = "LUH3")
}
\arguments{
\item{datasource}{Which LUH datasource should be used (\code{LUH2v2}, or \code{LUH3})}
}
\value{
A magpie object including the LUH total land area per cell
}
\description{
calculates the total land area included in a LUH data set
}
\author{
Patrick Rein
}
