% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticroppingOld.R
\name{calcMulticroppingOld}
\alias{calcMulticroppingOld}
\title{calcMulticroppingOld}
\usage{
calcMulticroppingOld(extend_future = FALSE, factortype = "CI")
}
\arguments{
\item{extend_future}{if TRUE}

\item{factortype}{CI: cropping intensity factor calculated as ratio of
    harvested to physical area where values above one
    indicate multicropping, below one fallow land (default)
MC: multiple cropping factor indicating areas that are
    harvested more than once in one year calculated taking
    fallow land into account explicitly:
    harvestedArea / (physicalArea - fallowLand)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates a multiple cropping factor based on area harvested,
             physical cropland area (and optionally fallow land). This function
             is an outdated version of calcMulticropping that is kept for
             Legacy reasons only for calcCropareaFAOLUH
}
\examples{
\dontrun{
calcOutput("MulticroppingOld")
}

}
\seealso{
[calcFAOLand()],
[calcCroparea()]
}
\author{
Benjamin Leon Bodirsky, David Chen, Felicitas Beier
}
