% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLPJmL.R
\name{convertLPJmL}
\alias{convertLPJmL}
\title{convertLPJmL}
\usage{
convertLPJmL(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\description{
converts units of data that was read in from LPJmL
             to units used in madrat and MAgPIE
}
\examples{
\dontrun{
readSource("LPJmL", convert = TRUE)
}

}
\seealso{
[readLPJmL()]
}
\author{
Felicitas Beier, Jan Philipp Dietrich
}
