% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLPJmL4.R
\name{convertLPJmL4}
\alias{convertLPJmL4}
\title{convertLPJmL4}
\usage{
convertLPJmL4(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Convert LPJmL content
}
\examples{
\dontrun{
readSource("LPJmL4", subtype = "soilc", convert = TRUE)
}

}
\seealso{
[readLPJmL()]
}
\author{
Kristine Karstens
}
