% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLPJmL4.R
\name{correctLPJmL4}
\alias{correctLPJmL4}
\title{correctLPJmL4}
\usage{
correctLPJmL4(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct LPJmL4 content
}
\examples{
\dontrun{
readSource("LPJmL4", subtype = "soilc", convert = "onlycorrect")
}

}
\seealso{
[correctLPJmL4()]
}
\author{
Kristine Karstens, Felicitas Beier
}
