% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLPJmLInputs_new.R
\name{correctLPJmLInputs_new}
\alias{correctLPJmLInputs_new}
\title{correctLPJmLInputs_new}
\usage{
correctLPJmLInputs_new(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\description{
correct LPJmLInputs_new content
}
\examples{
\dontrun{
readSource("LPJmLInputs_new", convert = "onlycorrect")
}

}
\author{
Felicitas Beier
}
