% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGAEZv4.R
\name{downloadGAEZv4}
\alias{downloadGAEZv4}
\title{downloadGAEZv4}
\usage{
downloadGAEZv4(subtype = NULL)
}
\arguments{
\item{subtype}{Name of file to be downloaded
Available are for example:
"mci_CRUTS32_Hist_0010.tif" (irrigated multiple cropping suitability for years 2000-2010),
"mcr_CRUTS32_Hist_0010.tif" (rainfed multiple cropping suitability for years 2000-2010)}
}
\description{
Download sub-set of GAEZ version 4 data
}
\examples{
\dontrun{
readSource("GAEZv4", convert = "onlycorrect")
}

}
\author{
Felicitas Beier
}
