% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadLPJmL.R
\name{downloadLPJmL}
\alias{downloadLPJmL}
\title{downloadLPJmL}
\usage{
downloadLPJmL(subtype = "lpjml5.10.0-m4:GSWP3-W5E5:historical:pnv:soilc")
}
\arguments{
\item{subtype}{Switch between different input
It consists of LPJmL version, climate model,
scenario and variable.}
}
\value{
metadata entry
}
\description{
Download LPJmL content by version, climate model and scenario
}
\examples{
\dontrun{
readSource("LPJmL", subtype = "lpjml5.10.0-m4:GSWP3-W5E5:historical:pnv:soilc", convert = FALSE)
}
}
\author{
Kristine Karstens, Marcos Alves, Felicitas Beier
}
