% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadLPJmLClimateInput.R
\name{downloadLPJmLClimateInput_new}
\alias{downloadLPJmLClimateInput_new}
\title{downloadLPJmLClimateInput_new}
\usage{
downloadLPJmLClimateInput_new(
  subtype = "ISIMIP3bv2:MRI-ESM2-0:ssp370:temperature"
)
}
\arguments{
\item{subtype}{Switch between different inputs (e.g. "ISIMIP3b:IPSL-CM6A-LR:historical:1850-2014:temperature")
Argument consists of GCM version, climate model, scenario and variable,
separated by ":"}
}
\value{
metadata entry
}
\description{
Download GCM climate input used for LPJmL runs
}
\examples{
\dontrun{
readSource("LPJmLClimateInput_new", convert = "onlycorrect")
}

}
\author{
Marcos Alves, Kristine Karstens
}
