% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlandcore-package.R
\docType{package}
\name{mrlandcore-package}
\alias{mrlandcore}
\alias{mrlandcore-package}
\title{mrlandcore: One-line description of this awesome package}
\description{
One-paragraph description of this awesome package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrlandcore}
}

}
\author{
\strong{Maintainer}: Kristine Karstens \email{karstens@pik-potsdam.de}

}
\keyword{internal}
