% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmL4.R
\name{readLPJmL4}
\alias{readLPJmL4}
\title{readLPJmL4}
\usage{
readLPJmL4(subtype = "LPJmL5:CRU4p02.soilc")
}
\arguments{
\item{subtype}{Switch between different input}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read LPJmL content
}
\examples{
\dontrun{
readSource("LPJmL4", subtype = "LPJmL5:CRU4p02.soilc", convert = "onlycorrect")
}

}
\author{
Kristine Karstens, Abhijeet Mishra, Felicitas Beier
}
