% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmLClimateInput_new.R
\name{readLPJmLClimateInput_new}
\alias{readLPJmLClimateInput_new}
\title{readLPJmLClimateInput_new}
\usage{
readLPJmLClimateInput_new(
  subset = "annualMean",
  subtype = "ISIMIP3bv2:MRI-ESM2-0:ssp370:temperature"
)
}
\arguments{
\item{subset}{Switch between different subsets of the same subtype
Available options are: "annualMean", "annualSum",
                       "monthlyMean", "monthlySum",
                       "wetDaysMonth"
Note that not all subtype-subset combinations make sense}

\item{subtype}{Switch between different inputs,
e.g. "ISIMIP3bv2:MRI-ESM2-0:ssp370:1850-2014:tas"
Available variables are: * tas -
                         * wet -
                         * per -}
}
\value{
MAgPIE objects with results on cellular level.
}
\description{
Read Climate data used as LPJmL inputs into MAgPIE objects
}
\examples{
\dontrun{
readSource("LPJmLClimateInput_new", subtype, convert = "onlycorrect")
}

}
\seealso{
\code{\link{readLPJmLClimateInput_new}}
}
\author{
Marcos Alves, Kristine Karstens, Felicitas Beier
}
