% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmLInputs_new.R
\name{readLPJmLInputs_new}
\alias{readLPJmLInputs_new}
\title{readLPJmLInputs_new}
\usage{
readLPJmLInputs_new(subtype = "lakeshare")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
This function reads in LPJmL inputs (inputs to LPJmL)
}
\examples{
\dontrun{
readSource("LPJmLInputs_new", subtype = "lakeshare", convert = FALSE)
}

}
\author{
Felicitas Beier
}
