% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCellArea.R
\name{toolCellArea}
\alias{toolCellArea}
\title{toolCellArea}
\usage{
toolCellArea(x)
}
\arguments{
\item{x}{corrdinated mag object}
}
\value{
mag object
}
\description{
Calculates  cell area for a given object
}
\author{
Kristine Karstens
}
