% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLPJmLCombine.R
\name{toolLPJmLCombine}
\alias{toolLPJmLCombine}
\title{toolLPJmLCombine}
\usage{
toolLPJmLCombine(lpjmlversion, climatetype, subtype, selectyears)
}
\arguments{
\item{lpjmlversion}{Switch between LPJmL versions (including add-ons (+*) for further version specification)}

\item{climatetype}{Switch between different climate scenarios}

\item{subtype}{Switch between runtype and variables in LPJmL}

\item{selectyears}{Subset of years}
}
\value{
configuration as list
}
\description{
Reads in and combines LPJmL data from LPJmL crops run.
             This function is used to select the correct runfolder specification
             and bind the objects from these runs together.
             Application example: rainfed and irrigated information comes from
             two different `crops` runs, since their growing periods differ.
             They have to be combined, e.g. for yields.
}
\author{
Felicitas Beier, Kristine Karstens
}
