% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLPJmLDefault.R
\name{toolLPJmLDefault}
\alias{toolLPJmLDefault}
\title{toolLPJmLDefault}
\usage{
toolLPJmLDefault(suppressNote = FALSE)
}
\arguments{
\item{suppressNote}{Suppress Notes}
}
\value{
configuration as list
}
\description{
Specify default settings for LPJmL version, climate and baseline settings
}
\author{
Kristine Karstens
}
