% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLPJmLHarmonization.R
\name{toolLPJmLHarmonization}
\alias{toolLPJmLHarmonization}
\title{toolLPJmLHarmonization}
\usage{
toolLPJmLHarmonization(version, climatetype)
}
\arguments{
\item{version}{Switch between LPJmL versions (including add-ons (+*) for further version specification)}

\item{climatetype}{Switch between different climate scenarios}
}
\value{
configuration as list
}
\description{
Specify default settings for LPJmL version and baseline settings
}
\author{
Kristine Karstens
}
