% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCroparea.R
\name{calcCroparea}
\alias{calcCroparea}
\title{calcCroparea}
\usage{
calcCroparea(
  sectoral = "kcr",
  physical = TRUE,
  cellular = FALSE,
  irrigation = FALSE,
  selectyears = "past",
  datasource = "LandInG"
)
}
\arguments{
\item{sectoral}{"kcr" MAgPIE items, and "lpj" LPJmL items}

\item{physical}{if TRUE the sum over all crops agrees with the cropland area per country}

\item{cellular}{if TRUE: calculates cellular MAgPIE crop area for all magpie croptypes.
Crop area from LUH3 crop types (c3ann, c4ann, c3per, c4per, cnfx)
are mapped to MAgpIE crop types using "FAO2LUH2MAG_croptypes" and doing
an intermediate step via harvested areas of FAO weight area within a
specific LUH crop type to divide into MAgPIE crop types.}

\item{irrigation}{If true: cellular areas are returned separated
into irrigated and rainfed (see setup in calcLUH3)}

\item{selectyears}{Vector of years to be returned}

\item{datasource}{Croparea data source (LandInG or FAOLUH).
Datasource FAOLUH is deprecated and is only kept for
backwards compatibility.}
}
\value{
areas of individual crops from FAOSTAT and weight
}
\description{
returns croparea
}
\author{
Felicitas Beier, Benjamin Leon Bodirsky
}
