% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFallow.R
\name{calcFallow}
\alias{calcFallow}
\title{calcFallow}
\usage{
calcFallow(cellular = FALSE, physical = TRUE)
}
\arguments{
\item{cellular}{TRUE for cellular outputs.}

\item{physical}{if TRUE it returns the physical area, with fallow of multicropping systems being
scaled to match physical areas of the rotation, if FALSE the area harvested
being fallow is returned}
}
\value{
MAgPIE object containing fallow land in Mha
}
\description{
Returns fallow land of our default datasource.
The estimates are scaled to match our Land classes in the Croparea function.
Due to multiple cropping, harvested cropland area can be greater than non-fallow land area
and even greater than physical cropland area.
}
\examples{
\dontrun{
calcOutput("Fallow")
}
}
\seealso{
\code{\link{readLandInG}}
}
\author{
David Hoetten, Felicitas Beier
}
