% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLanduseInitialisation.R
\name{calcLanduseInitialisation}
\alias{calcLanduseInitialisation}
\title{calcLanduseInitialisation}
\usage{
calcLanduseInitialisation(
  cellular = FALSE,
  nclasses = "seven",
  cells = "lpjcell",
  selectyears = "past_til2020",
  input_magpie = FALSE
)
}
\arguments{
\item{cellular}{cellular (TRUE) or country-level/regional (FALSE) data?
For country-level vs regional data: remember to set "aggregate" to FALSE.}

\item{nclasses}{options are either "six", "seven" or "nine".
\itemize{
\item "five" aggregates forestry and forest, includes "crop", "past", "forest", "urban" and "other"
\item "six" includes the original land use classes "crop", "past", "forestry", "forest", "urban" and "other"
\item "seven" separates primary and secondary forest and includes "crop", "past", "forestry", "primforest",
"secdforest", "urban" and "other"
\item "nine" adds the separation of pasture and rangelands, as well as a differentiation of primary
and secondary non-forest vegetation and therefore returns "crop", "past", "range", "forestry", "primforest",
"secdforest", "urban", "primother" and "secdother"
}}

\item{cells}{if cellular is TRUE: "magpiecell" for 59199 cells or "lpjcell" for 67420 cells}

\item{selectyears}{default on "past"}

\item{input_magpie}{applies area fix (set cells with zero area to minimal value to
not disturb aggregating to clusters)}
}
\value{
List of magpie object with results on country or cellular level, weight on cellular level,
unit and description.
}
\description{
Calculates the cellular MAgPIE landuse initialisation area.
             Most of the calculation happens in calcLanduseInitialisationBase.
}
\examples{
\dontrun{
calcOutput("LanduseInitialisation")
}
}
\author{
Jan Philipp Dietrich, Benjamin Leon Bodirsky, Kristine Karstens, Felicitas Beier, Patrick v. Jeetze
}
