% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmLharmonize.R
\name{calcLPJmLharmonize}
\alias{calcLPJmLharmonize}
\title{calcLPJmLharmonize}
\usage{
calcLPJmLharmonize(
  version = "lpjml5.9.5-m1",
  climatetype = "MRI-ESM2-0:ssp370",
  subtype = "pnv:soilc",
  subdata = NULL,
  stage = "harmonizedScenario"
)
}
\arguments{
\item{version}{Switch between LPJmL versions (including addons for further version specification)}

\item{climatetype}{Switch between different climate scenarios}

\item{subtype}{Switch between different LPJmL output variables as specified in readLPJmL}

\item{subdata}{Switch between data dimension subitems
#### Kristine (To Do): explain switch better (maybe give examples). For me the use case wasn't clear.}

\item{stage}{degree of processing: harmonizedHistorical (until reference year of historical data)
                      harmonizedScenario (until reference year for harmonized Scenarios)
See toolLPJmLHarmonization for current settings.}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Handle harmonization of LPJmL data
}
\examples{
\dontrun{
calcOutput("LPJmLharmonize", subtype = "pnv:soilc", aggregate = FALSE)
}
}
\seealso{
[toolLPJmLHarmonization()]
}
\author{
Kristine Karstens, Felicitas Beier
}
