% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmLharmonize.R
\name{calcLPJmLharmonize}
\alias{calcLPJmLharmonize}
\title{calcLPJmLharmonize}
\usage{
calcLPJmLharmonize(
  lpjmlversion = "lpjml5.9.5-m1",
  climatetype = "MRI-ESM2-0:ssp370",
  subtype = "pnv:soilc",
  subdata = NULL
)
}
\arguments{
\item{lpjmlversion}{Switch between LPJmL versions (including addons for further version specification)}

\item{climatetype}{Switch between different climate scenarios}

\item{subtype}{Switch between different LPJmL output variables as specified in readLPJmL}

\item{subdata}{Selection of subitems of object.
This argument can be used to split up the data in smaller objects
where only a sub-set of the data is needed or
for better handling where otherwise memory issues would occur due to the object size.}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Handle harmonization of LPJmL data
}
\examples{
\dontrun{
calcOutput("LPJmLharmonize", subtype = "pnv:soilc", aggregate = FALSE)
}
}
\seealso{
[toolLPJmLHarmonization()]
}
\author{
Kristine Karstens, Felicitas Beier
}
