% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLUH3.R
\name{calcLUH3}
\alias{calcLUH3}
\title{calcLUH3}
\usage{
calcLUH3(
  landuse_types = "magpie",
  irrigation = FALSE,
  cellular = FALSE,
  cells = "lpjcell",
  yrs = seq(1965, 2020, 5)
)
}
\arguments{
\item{landuse_types}{magpie: magpie landuse classes,
LUH3: original landuse classes
flooded: flooded areas as reported by LUH}

\item{irrigation}{if true: areas are returned separated by irrigated and rainfed,
if false: total areas}

\item{cellular}{if true: dataset is returned on 0.5 degree resolution}

\item{cells}{Switch between "magpiecell" (59199) and "lpjcell" (67420) (default: "lpjcell")
NOTE: This setting also affects the sums on country level!}

\item{yrs}{years to be returned (default: seq(1965, 2020, 5))}
}
\value{
List of magpie objects with results on country level,
        weight on country level, unit and description
}
\description{
Integrates the LUH3 landuse-dataset
}
\examples{
\dontrun{
calcOutput("LUH3")
}
}
\seealso{
[calcLanduseInitialisation()]
}
\author{
Wanderson Costa, Pascal Sauer, Miodrag Stevanovic, Alexandre Koberle
}
