% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticropping.R
\name{calcMulticropping}
\alias{calcMulticropping}
\title{calcMulticropping}
\usage{
calcMulticropping()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Multicropping factor, telling how many harvests are produced
on one unit of phyiscal area. Fallow lands are not counted in, so the factor
is always larger than 1.
}
\examples{
\dontrun{
calcOutput("Multicropping")
}

}
\seealso{
[calcCroparea()]
}
\author{
Benjamin Leon Bodirsky
}
