% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNPPyearly.R
\name{calcNPPyearly}
\alias{calcNPPyearly}
\title{calcNPPyearly}
\usage{
calcNPPyearly(
  subtype = "preind",
  lpjml = "lpjml5.10.0-m1",
  climatetype = "MRI-ESM2-0:ssp370",
  unit = "tC/m2"
)
}
\arguments{
\item{subtype}{Type of NPP to be returned.
Options:
"preind": reference pre-industrial NPP from pi-control run
"pnv": potetnial NPP from potential natural vegetation run}

\item{lpjml}{Defines LPJmL version for main crop inputs}

\item{climatetype}{Switch between different climate scenarios}

\item{unit}{Unit to be returned.
Options:
"tC/m2" (tons carbon per square meter, original LPJmL unit)
"tC/ha" (tons carbon per hectare)
"tDM/ha" (tons dry matter per hectare, standard MAgPIE unit)}
}
\value{
magpie object in cellular resolution
}
\description{
This function extracts NPP from LPJmL and transforms it to
             yearly values including unit transformations
}
\examples{
\dontrun{
calcOutput("NPPyearly", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
