% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLPJmL.R
\name{correctLPJmL}
\alias{correctLPJmL}
\title{correctLPJmL}
\usage{
correctLPJmL(x)
}
\arguments{
\item{x}{list of magpie object and unit provided by the read function}
}
\description{
corrects data that was read in from LPJmL and converts units
             to MAgPIE standard
}
\examples{
\dontrun{
readSource("LPJmL", convert = "onlycorrect")
}

}
\seealso{
[readLPJmL()]
}
\author{
Felicitas Beier, Kristine Karstens
}
