% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrlandcore-package.R
\docType{package}
\name{mrlandcore-package}
\alias{mrlandcore}
\alias{mrlandcore-package}
\title{mrlandcore: One-line description of this awesome package}
\description{
One-paragraph description of this awesome package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/mrlandcore}
}

}
\author{
\strong{Maintainer}: Felicitas Beier \email{beier@pik-potsdam.de}

Authors:
\itemize{
  \item Kristine Karstens
  \item Marcos Alves
  \item Jan Philipp Dietrich
  \item Benjamin Leon Bodirsky
  \item David Hoetten
  \item Florian Humpenoeder
  \item Jens Heinke
  \item Patrick v. Jeetze
  \item Abhijeet Mishra
  \item Felcitas Beier
  \item Stephen Wirth
  \item David Chen
  \item Ulrich Kreidenweis
}

}
\keyword{internal}
