% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCellArea.R
\name{toolCellArea}
\alias{toolCellArea}
\title{toolCellArea}
\usage{
toolCellArea(x)
}
\arguments{
\item{x}{magpie object with coordinate data}
}
\value{
mag object
}
\description{
Calculates cell area (in Mha) for a given object
}
\author{
Kristine Karstens, Felicitas Beier
}
