% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolLPJmLScenario.R
\name{toolLPJmLScenario}
\alias{toolLPJmLScenario}
\title{toolLPJmLScenario}
\usage{
toolLPJmLScenario(version, climatetype, subtype)
}
\arguments{
\item{version}{Switch between LPJmL versions (including add-ons (+*) for further version specification)}

\item{climatetype}{Switch between different climate scenarios}

\item{subtype}{Switch between runtype and variables in LPJmL}
}
\value{
configuration as list
}
\description{
Specify default settings for LPJmL version and baseline settings
}
\author{
Kristine Karstens
}
